# backend/docs/swagger.yaml
definitions:
  repo.Client:
    properties:
      address_id:
        type: string
      address_robust:
        type: string
      bandwidth:
        type: string
      data_cap:
        type: number
      location:
        type: string
      monthly_storage:
        type: string
      name:
        type: string
      service_requirement:
        type: string
      storage_capacity:
        type: number
      storage_deals:
        type: integer
      total_data_cap:
        type: number
      use_case:
        type: string
      used_data_cap:
        type: number
      verified_deals:
        type: integer
    type: object
  repo.ClientHistoryDealStatsItem:
    properties:
      avg_price:
        type: number
      avg_verified_price:
        type: number
      data_cap:
        type: number
      first_deal_time:
        type: integer
      iso_code:
        type: string
      max_piece_size:
        type: number
      min_piece_size:
        type: number
      name:
        type: string
      owner:
        type: string
      provider:
        type: string
      quality_adj_power:
        type: number
      raw_power:
        type: number
      region:
        type: string
      reputation_score:
        type: integer
      retrieval_success_rate:
        type: number
      review_score:
        type: number
      reviews:
        type: integer
      storage_capacity:
        type: number
      storage_deals:
        description: relevant client deal
        type: integer
      storage_success_rate:
        type: number
    type: object
  repo.ClientInfo:
    properties:
      address_id:
        type: string
      bandwidth:
        type: string
      data_cap:
        type: number
      monthly_storage:
        type: string
      service_requirement:
        type: string
      storage_capacity:
        type: number
      storage_deals:
        type: integer
      total_data_cap:
        type: number
      use_case:
        type: string
      used_data_cap:
        type: number
      verified_deals:
        type: integer
    type: object
  repo.ClientProfile:
    properties:
      address_id:
        type: string
      address_robust:
        type: string
      avatar:
        type: string
      contact_email:
        type: string
      description:
        type: string
      github:
        type: string
      location:
        type: string
      logo:
        type: string
      name:
        type: string
      slack:
        type: string
      twitter:
        type: string
      type:
        type: string
      uid:
        type: string
    type: object
...
  repo.Provider:
    properties:
      address:
        type: string
      available_deals:
        type: string
      bandwidth:
        type: string
      certification:
        type: string
      experience:
        type: string
      first_deal_time:
        type: integer
      is_member:
        type: string
      iso_code:
        type: string
      max_piece_size:
        type: number
      min_piece_size:
        type: number
      name:
        type: string
      owner:
        type: string
      parallel_deals:
        type: string
      price:
        type: string
      quality_adj_power:
        type: number
      raw_power:
        type: number
      region:
        type: string
      renewable_energy:
        type: string
      reputation_score:
        type: integer
      retrieval_success_rate:
        type: number
      review_score:
        type: number
      reviews:
        type: integer
      sealing_speed:
        type: string
      storage_deals:
        type: integer
      storage_success_rate:
        type: number
      verified_price:
        type: string
    type: object
  repo.Review:
    properties:
      client:
        type: string
      content:
        type: string
      created_at:
        $ref: '#/definitions/types.UnixTime'
      provider:
        type: string
      score:
        type: integer
      title:
        type: string
    type: object
  repo.SPOwnerProfile:
    properties:
      address_id:
        type: string
      address_robust:
        type: string
      avatar:
        type: string
      contact_email:
        type: string
      description:
        type: string
      github:
        type: string
      location:
        type: string
      logo:
        type: string
      name:
        type: string
      reputation_score:
        type: number
      review_score:
        type: number
      reviews:
        type: integer
      slack:
        type: string
      twitter:
        type: string
      type:
        type: string
      uid:
        type: string
    type: object
...
  server.ClientDetailParams:
    properties:
      bandwidth:
        example: 300M
        maxLength: 128
        type: string
      monthly_storage:
        example: 10TiB
        maxLength: 128
        type: string
      service_requirement:
        example: More information about us.
        maxLength: 1024
        type: string
      use_case:
        example: Entertainment/Media/Science
        maxLength: 128
        type: string
    type: object
...
  server.ReviewParams:
    properties:
      content:
        example: Great. It's the ultimate experience.
        maxLength: 1024
        type: string
      provider:
        example: f01662887
        type: string
      score:
        example: 5
        maximum: 5
        minimum: 0
        type: integer
      title:
        example: Ultimate Experience
        maxLength: 128
        type: string
    required:
    - content
    - provider
    - score
    - title
    type: object
...
  types.UnixTime:
    properties:
      time.Time:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /chat/history:
    get:
      consumes:
      - application/json
      parameters:
      - default: Bearer YOUR_JWT
        description: jwt access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Partner uid
        in: query
        name: partner
        type: string
      - description: Max records to return
        in: query
        maximum: 100
        name: limit
        type: integer
      - description: Option, unix time(seconds). It returns last messages if you not
          set.
        in: query
        name: before
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Get chat history
      tags:
      - chat
  /client/detail:
    post:
      consumes:
      - application/json
      parameters:
      - default: Bearer YOUR_JWT
        description: jwt access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: client detail param
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/server.ClientDetailParams'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: client modify detailed information
      tags:
      - client
  /client/review:
    post:
      consumes:
      - application/json
      parameters:
      - default: Bearer YOUR_JWT
        description: jwt access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: review param
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/server.ReviewParams'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Submit a review
      tags:
      - client
  /clientDetail:
    get:
      consumes:
      - application/json
      parameters:
      - description: Client address id.
        example: f01624861
        in: query
        name: address_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.ClientInfo'
      summary: Get client detailed info
      tags:
      - public
  /clientHistoryDealStats:
    get:
      consumes:
      - application/json
      parameters:
      - description: Which page to return
        in: query
        name: page
        type: integer
      - description: Max records to return
        in: query
        maximum: 100
        name: page_size
        type: integer
      - description: Client address id.
        example: f01624861
        in: query
        name: address_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.ClientHistoryDealStatsItem'
            type: array
      summary: Get client statistics about history deals
      tags:
      - public
  /clientProfile:
    get:
      consumes:
      - application/json
      parameters:
      - description: Client address id.
        example: f01624861
        in: query
        name: address_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.ClientProfile'
      summary: Get client profile
      tags:
      - public
  /clientReviews:
    get:
      consumes:
      - application/json
      parameters:
      - description: Which page to return
        in: query
        name: page
        type: integer
      - description: Max records to return
        in: query
        maximum: 100
        name: page_size
        type: integer
      - description: Client address id.
        example: f01624861
        in: query
        name: address_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.Review'
            type: array
      summary: Get client reviews
      tags:
      - public
  /clients:
    get:
      consumes:
      - application/json
      parameters:
      - description: Which page to return
        in: query
        name: page
        type: integer
      - description: Max records to return
        in: query
        maximum: 100
        name: page_size
        type: integer
      - default: data_cap
        description: 'Sorting option. Example: sort_by=data_cap'
        enum:
        - storage_deals
        - storage_capacity
        - total_data_cap
        - used_data_cap
        - data_cap
        - verified_deals
        in: query
        name: sort_by
        type: string
      - default: desc
        description: 'Option to order providers. Example: order=desc'
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - description: 'Search client by keyword, support client id/name/location. Example:
          f01234'
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.Client'
            type: array
      summary: Client list
      tags:
      - public
...
  /spOwnerProfile:
    get:
      consumes:
      - application/json
      parameters:
      - description: Owner address id.
        example: f01694606
        in: query
        name: address_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.SPOwnerProfile'
      summary: Get SP owner profile
      tags:
      - public
  /spOwnerReviews:
    get:
      consumes:
      - application/json
      parameters:
      - description: Which page to return
        in: query
        name: page
        type: integer
      - description: Max records to return
        in: query
        maximum: 100
        name: page_size
        type: integer
      - description: Owner address id.
        example: f01694606
        in: query
        name: address_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.Review'
            type: array
      summary: Get reviews by SP owner
      tags:
      - public
  /spServiceDetail:
    get:
      consumes:
      - application/json
      parameters:
      - description: Owner address id.
        example: f01694606
        in: query
        name: address_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.Provider'
            type: array
      summary: Get SP service detail list by owner
      tags:
      - public
...
swagger: "2.0"